% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{str_extract_between}
\alias{str_extract_between}
\title{Extract characters in a string which occur after a specified pattern, and before a (different) specified pattern}
\usage{
str_extract_between(string, pattern1, pattern2)
}
\arguments{
\item{string}{The string from which to extract.}

\item{pattern1}{Pattern to look for before the return content of a string}

\item{pattern2}{Pattern to look for after the return content of a string

The default interpretation is a regular expression, as described in stringi::stringi-search-regex. Control options with regex().

Match a fixed string (i.e. by comparing only bytes), using fixed(). This is fast, but approximate. Generally, for matching human text, you'll want coll() which respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with boundary(). An empty pattern, "", is equivalent to boundary("character").}
}
\value{
The part of the string after the first pattern and before the second pattern.
}
\description{
Vectorised over `string` and `pattern`.
}
\examples{
test_string <- 'url.com/string_before_pattern_after-another_string_with_pattern2/'
str_extract_between(string = test_string, pattern1 = "_pattern_", pattern2 = "_pattern2")


}
