% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_context.R
\name{str_extract_context}
\alias{str_extract_context}
\title{Extract pattern from a string along with characters that occur before and after the pattern.}
\usage{
str_extract_context(string, pattern, window_size = 10)
}
\arguments{
\item{string}{The string from which to extract.}

\item{pattern}{Pattern to look for.}

\item{window_size}{Number of characters to return before and after the pattern.}
}
\value{
A character vector with pattern matches and their context.
}
\description{
#' Vectorised over `string`, `pattern`, and `window_size`.
}
\examples{
test_string <- 'To be or not to be, that is the question.'
str_extract_context(test_string, pattern = 'be', window_size = 4)

}
